[feat] split converter
